requirements.txt
requests
pulp
python-dotenv
beautifulsoup4
lxml
pandas
pyarrow
